planner/mtp/requirements.txt
requests
fastapi
uvicorn[standard]
confluent-kafka
python-dotenv